Clean up more types
